v1.2.2 display sample response for wsdl
Functionality added, program now adds a sample generated response for a
wsdl when the wsdl is loaded.
Change, author has been removed from the about page.
Change, the update dialog has been removed and is included in the about
page.
Bug fix, bug introduced in v1.2.1 license file is now read in and
converted to the correct line ending for windows or linux